Rename package to `oemof.pg`
This entails changing the name of the PyPi- as well as the
import-package. The latter means that the directory structure and
consequently `setup`'s `package_dir` option needed to be changed.
All packages are still there, expect that you now import them from
`oemof.db` instead of `oemof_pg`. In order to not create an
`oemof.db.db` package, I put what previously was `oemof_pg.db` into
`oemof.db` directly so for that one the imports change as follows:

  - `import oemof_pg.db`           becomes `import oemof.db` and
  - `from oemof_pg.db import NAME` becomes `from oemof.db import NAME`.

In accordance with the `setuptools` documentation on namespace packages,
I also supplied the `namespace_package` option.

I put the description update into this commit too, because I think it's
a closely related change to the package renaming.