Fix `config.ini` documentation reference
I guess there shouldn't have been a space between ":download:" and
"`config.py`" and that the `config` module was meant so I wanted to
reference that. Instead I ended up linking to the `README`, as the
structure and location of `congig.ini` is described pretty clearly
there. In order to do that I had to add an explicit [label][0] to
the `Configuration` section of the `README` as I couldn't get
hyperlinking to work properly otherwise.

[0]: http://www.sphinx-doc.org/en/stable/markup/inline.html#cross-referencing-arbitrary-locations