Merge generated documentation with existing one
This entailed a lot of things, but I guess the most complicated thing
was merging the `conf.py` files. I threw out most of the unused
commented stuff in the old `conf.py` because I found it rather
distracting. And since I might need some of the options from the old
`conf.py` file, I kept a backup of it in `conf.py.old`.